Nice java package convention derrived
