--- aem.app.common/src/main/kotlin/com/company/aem/example/common/JsonUtils.kt
@@ -1,9 +1,0 @@
-package com.company.aem.example.common
-
-import com.google.gson.GsonBuilder
-
-object JsonUtils {
-
-  val GSON = GsonBuilder().serializeNulls().setPrettyPrinting().create()
-
-}
