+++ aem.app.common/src/main/kotlin/com/company/example/aem/app/common/AppUtils.kt
@@ -1,0 +1,16 @@
+package com.company.example.aem.app.common
+
+import com.google.common.base.Charsets
+import java.util.*
+
+object AppUtils {
+
+  const val APP_NAME = "Example"
+
+  const val LABEL_PREFIX = APP_NAME
+
+  val CHARSET = Charsets.UTF_8
+
+  val TIMEZONE = TimeZone.getTimeZone("GMT")
+
+}
