+++ aem.app.common/src/main/kotlin/com/company/example/aem/app/common/JsonUtils.kt
@@ -1,0 +1,9 @@
+package com.company.example.aem.app.common
+
+import com.google.gson.GsonBuilder
+
+object JsonUtils {
+
+  val GSON = GsonBuilder().serializeNulls().setPrettyPrinting().create()
+
+}
