+++ aem.app.core/src/main/content/jcr_root/apps/example/core/components/home-page/home-page.html
@@ -1,4 +1,4 @@
-<sly data-sly-use.page="com.company.aem.example.core.components.page.PageModel"/>
+<sly data-sly-use.page="com.company.example.aem.app.core.components.page.PageModel"/>
 <sly data-sly-use.clientLib="${'/libs/granite/sightly/templates/clientlib.html'}"/>
 <!DOCTYPE html>
 <html lang="${page.languageCode}">
