+++ aem.app.core/src/main/kotlin/com/company/example/aem/app/core/services/posts/Post.kt
@@ -1,0 +1,8 @@
+package com.company.example.aem.app.core.services.posts
+
+data class Post(
+  val userId: Int,
+  val id: Int,
+  val title: String,
+  val body: String
+)
