+++ aem.migration/src/main/content/META-INF/vault/filter.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <workspaceFilter version="1.0">
-    <filter root="/etc/groovyconsole/scripts/example"/>
+    <filter root="/etc/groovyconsole/scripts/aecu/example"/>
 </workspaceFilter>
