+++ aem.migration/src/main/content/jcr_root/etc/groovyconsole/scripts/aecu/example/migrate.groovy
